+++ src/tqec/__init__.py
@@ -1,29 +1,17 @@
 from . import (
-    detectors,
-    generation,
+    circuit,
     noise_models,
     plaquette,
     templates,
 )
 from ._version import __version__
-from .detectors import (
-    RelativeMeasurementData,
-    make_detector,
-    make_observable,
-    make_shift_coords,
-    transform_to_stimcirq_compatible,
-)
-from .display import (
-    display_template,
-    display_templates_svg,
-)
+from .circuit import generate_circuit
 from .enums import (
     CornerPositionEnum,
     PlaquetteOrientation,
     TemplateRelativePositionEnum,
 )
 from .exceptions import TQECException
-from .generation import generate_circuit
 from .noise_models import (
     AfterCliffordDepolarizingNoise,
     AfterResetFlipNoise,
@@ -60,4 +48,6 @@ from .templates import (
     ScalableCorner,
     Template,
     TemplateWithIndices,
+    display_template,
+    display_templates_svg,
 )
