+++ src/tqec/circuit/__init__.py
@@ -1,0 +1,8 @@
+from .circuit import generate_circuit
+from .detectors import (
+    RelativeMeasurementData,
+    make_detector,
+    make_observable,
+    make_shift_coords,
+    transform_to_stimcirq_compatible,
+)
