+++ src/tqec/circuit/detectors/__init__.py
