+++ src/tqec/circuit/detectors/operation.py
