+++ src/tqec/circuit/schedule.py
@@ -5,7 +5,8 @@ import typing
 from copy import deepcopy
 
 import cirq
-from tqec.detectors.operation import Detector, make_detector
+
+from tqec.circuit.detectors.operation import Detector, make_detector
 from tqec.exceptions import TQECException
