--- src/tqec/generation/__init__.py
@@ -1,1 +1,0 @@
-from .circuit import generate_circuit
