+++ src/tqec/plaquette/library/initialisation.py
@@ -5,7 +5,7 @@ import typing as ty
 import cirq
 from tqec.enums import PlaquetteOrientation
 from tqec.plaquette.plaquette import RoundedPlaquette, SquarePlaquette
-from tqec.plaquette.schedule import ScheduledCircuit
+from tqec.circuit.schedule import ScheduledCircuit
 
 
 class ZSquareInitialisationPlaquette(SquarePlaquette):
