+++ src/tqec/plaquette/library/xxxx.py
@@ -5,7 +5,7 @@ import cirq
 from tqec.plaquette.library.utils.detectors import make_memory_experiment_detector
 from tqec.plaquette.library.utils.pauli import make_pauli_syndrome_measurement_circuit
 from tqec.plaquette.plaquette import SquarePlaquette
-from tqec.plaquette.schedule import ScheduledCircuit
+from tqec.circuit.schedule import ScheduledCircuit
 
 
 class XXXXMemoryPlaquette(SquarePlaquette):
