+++ src/tqec/templates/display.py
